Add and bind `ctrlf-forward-symbol-at-point`
This allows users that are accustomed to
`isearch-forward-symbol-at-point` semantics to use ctrlf with ease. The
`M-s .` mapping to start a symbol at point search (or populate an
existing search with symbol at point) now work, in addition to the
existing `M-n` option to use the first history element (which is also
the symbol at point).