Allow `ctrlf--message` to be used any time
In order to allow ctrlf commands to be able to display error messages
when the search does not start, `ctrlf--message` needs to defer to
`minibuffer-message` when the `ctrlf--minibuffer` is `nil`. This also
means we need to set it to `nil` when cleaning up.

This function had to be re-indented and a line break added.