More like isearch (#48)
* Add isearch-like bindings for case and style toggles

* Add optional `arg` to `ctrlf--start` for minibuffer initial contents

This enables wrapper functions to be created with starting search terms.

* Add optional `arg` to `ctrlf-forward` too

With the optional `arg`, the search will start with (or continue with)
it as the given search term. This lets wrapper functions supply initial
search terms.

* Add and bind `ctrlf-forward-symbol-at-point`

This allows users that are accustomed to
`isearch-forward-symbol-at-point` semantics to use ctrlf with ease. The
`M-s .` mapping to start a symbol at point search (or populate an
existing search with symbol at point) now work, in addition to the
existing `M-n` option to use the first history element (which is also
the symbol at point).

* Implement `ctrlf-occur` which acts much like `isearch-occur`

This is surprisingly straightforward because of the existing logic to
automatically translate the input based on its style into a regexp.

* Add optional `pos` to `ctrlf--start` as starting point

This allows us to start the search from another position (say, the
beginning of the symbol at point), but retain the user's actual starting
position so that `ctrlf-cancel` behaves appropriately.

* Improve handling of current point in `ctrlf-forward-symbol-at-point`

This now makes the search start at the beginning of the symbol at point,
so that if the point is within a symbol, it becomes the first search
result. Canceling the search drops the user back at the appropriate
position, the point at which they started.

* Add symbol boundary constructs

We may want to make this optional. The former behavior, while less
consistent with isearch, was actually quite nice. It would search for
the symbol at point, but include strings of which it was the prefix, and
these could then be added to the search with `M-s .`.

* Allow `ctrlf--message` to be used any time

In order to allow ctrlf commands to be able to display error messages
when the search does not start, `ctrlf--message` needs to defer to
`minibuffer-message` when the `ctrlf--minibuffer` is `nil`. This also
means we need to set it to `nil` when cleaning up.

This function had to be re-indented and a line break added.

* Improve handling of no symbol at point

We now skip the search if there is no symbol at point, and display an
error message. This removes the workaround of sending an empty string to
`ctrlf-forward` when there was no symbol.

* Add `ctrlf-translate-symbol` translator

This simple refactor allows us to provide `ctrlf-forward-symbol` and
`ctrlf-backward-symbol`, and remap their isearch counterparts.

Furthermore, `ctrlf-forward-symbol-at-point` no longer has this logic
embedded in it.

* Add optional `pos` to `ctrlf-forward` as starting point

This is simply passed to `ctrlf--start` if the search is being started,
which allows wrapper functions to not deal with this special case.

* Add `ctrlf-toggle-regexp` and `ctrlf-toggle-symbol`

So that we can correctly bind `M-s r` and `M-s _` when in a CTRLF
session.

* Update readme and changelog

* Don't `ctrlf-forward-symbol-at-point` twice

We're already defining it globally (as Isearch does) and it doesn't need
to be rebound again here. The rest of these, however, are only bound
during a search.

* Simplify setting the current starting point in symbol search

We can avoid moving `point` by setting the _current_ starting point
instead of the "starting point," reducing complexity.

* Rename `arg` and `pos` to `initial-contents` and `position`